+++ frontend/web-front/src/Signup.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+import ReactDOM from "react-dom";
+
+export const Signup: React.VFC = () => {
+  return <>Signup</>;
+};
+export default Signup;
